pyproject.toml
[tool.poetry]
name = "weather-map"
version = "0.1.0"
description = ""
authors = ["roger <r1990227@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
pre-commit = "^4.0.1"
pyowm = "^3.3.0"
pytz = "^2024.2"
python-dateutil = "^2.9.0.post0"
django-tailwind = {extras = ["reload"], version = "^3.8.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
